+++ src/game_state.rs
@@ -18,8 +18,8 @@ pub struct CardPile(pub Vec<Card>);
 pub struct GameState {
     pub deck: Vec<Card>,
     pub deck_drawn: Vec<Card>,
-    pub columns: [CardColumn; Self::COLUMN_COUNT as usize],
-    pub card_piles: [CardPile; Self::CARD_PILES_COUNT as usize],
+    pub columns: [CardColumn; Self::COLUMN_COUNT],
+    pub card_piles: [CardPile; Self::CARD_PILES_COUNT],
 }
 
 pub trait CardCollection {
@@ -127,12 +127,12 @@ impl CardColumn {
 }
 
 impl GameState {
-    pub const COLUMN_COUNT: u8 = 7;
-    pub const CARD_PILES_COUNT: u8 = 4;
+    pub const COLUMN_COUNT: usize = 7;
+    pub const CARD_PILES_COUNT: usize = 4;
 
     pub fn init(mut deck: Vec<Card>) -> Self {
-        let mut columns: [CardColumn; Self::COLUMN_COUNT as usize] = Default::default();
-        let card_piles: [CardPile; Self::CARD_PILES_COUNT as usize] = Default::default();
+        let mut columns: [CardColumn; Self::COLUMN_COUNT] = Default::default();
+        let card_piles: [CardPile; Self::CARD_PILES_COUNT] = Default::default();
 
         for (i, column) in columns.iter_mut().enumerate() {
             for _ in 0..=i {
@@ -170,7 +170,7 @@ impl GameState {
     }
 
     pub fn victory() -> Self {
-        let mut card_piles: [CardPile; Self::CARD_PILES_COUNT as usize] = Default::default();
+        let mut card_piles: [CardPile; Self::CARD_PILES_COUNT] = Default::default();
 
         for (index, suit) in Suit::iter().enumerate() {
             for rank in Rank::iter() {
